feat: refactor ActionButtons, CountdownEdit, and ToolButtons components; add custom hooks for action and countdown logic
